try fix windows ci x86

+++ test/runtests.jl
@@ -129,7 +129,7 @@ import CpuId
 
     @testset "CPU core kinds" begin
         @test num_cpukinds() >= 1
-        @test typeof(num_virtual_cores_cpukinds()) in (Nothing, Vector{Int64})
+        @test typeof(num_virtual_cores_cpukinds()) in (Nothing, Vector{Int})
         if !isnothing(num_virtual_cores_cpukinds())
             @test all(>=(1), num_virtual_cores_cpukinds())
             @test sum(num_virtual_cores_cpukinds()) == num_virtual_cores()
